chore: place routes in another file
